\name{PublicEmploymentAnalysis-package}
\alias{PublicEmploymentAnalysis-package}
\alias{PublicEmploymentAnalysis}
\docType{package}
\title{
\packageTitle{PublicEmploymentAnalysis}
}
\description{
\packageDescription{PublicEmploymentAnalysis}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{PublicEmploymentAnalysis}
\packageIndices{PublicEmploymentAnalysis}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{PublicEmploymentAnalysis}

Maintainer: \packageMaintainer{PublicEmploymentAnalysis}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R ~~
~~ documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
%% ~~ simple examples of the most important functions ~~
}
